+++ src/parse/semantic/func.rs
@@ -10,16 +10,35 @@ fn make_init_args_node(
         next: match last_node {
             Some(node) => Some(Box::new(node.clone())),
             None => None
-        }
+        },
+        multiple: None,
     }
 }
 
 //  関数の引数
-fn make_args_node(tokens: &Vec<token::Token>, index: &mut usize) -> manager::func::FuncArgsNode {
+fn make_args_node(
+        tokens: &Vec<token::Token>,
+        index: &mut usize
+) -> Result<manager::func::FuncArgsNode, err_kind::ErrorsKind> {
     let mut args_node = make_init_args_node(&None);
+    let mut var_status: Option<manager::variable::MultipleVar> = None;
 
     while tokens.len() > *index {
         match tokens[*index].kind {
+            token::TokenKind::TokenVarMut => {
+                if var_status == None {
+                    var_status = Some(manager::variable::MultipleVar::IsMut);
+                } else {
+                    Err(err_kind::ErrorsKind::MultipleMutabilitySpecifiers)?;
+                }
+            }
+            token::TokenKind::TokenVarImm => {
+                if var_status == None {
+                    var_status = Some(manager::variable::MultipleVar::IsImm);
+                } else {
+                    Err(err_kind::ErrorsKind::MultipleMutabilitySpecifiers)?;
+                }
+            }
             token::TokenKind::TokenName => args_node.name = tokens[*index].lexeme.clone(),
             token::TokenKind::TokenSpace => {},
             token::TokenKind::TokenRParen => {
@@ -35,20 +54,26 @@ fn make_args_node(tokens: &Vec<token::Token>, index: &mut usize) -> manager::fun
                         ).unwrap().value.clone()
                     )
                 };
+                args_node.multiple = var_status.clone();
                 continue;
             }
-            token::TokenKind::TokenComma =>
-                args_node = make_init_args_node(&Some(args_node)),
+            token::TokenKind::TokenComma => {
+                args_node = make_init_args_node(&Some(args_node));
+                var_status = None;
+            }
             _ => {}
         }
         *index += 1;
     }
 
-    args_node
+    Ok(args_node)
 }
 
 //  関数ヘッダーを作成する関数
-pub fn make_func_header(tokens: Vec<token::Token>, index: &mut usize) -> manager::func::FuncNode {
+pub fn make_func_header(
+        tokens: Vec<token::Token>,
+        index: &mut usize
+) -> Result<manager::func::FuncNode, err_kind::ErrorsKind> {
     let mut func_start_keyword: bool = false;
     //  関数の名前がある場所を代入
     let mut func_name_index: Option<usize> = None;
@@ -73,11 +98,11 @@ pub fn make_func_header(tokens: Vec<token::Token>, index: &mut usize) -> manager
                 break;
             }
             token::TokenKind::TokenLParen => {
-                args = make_args_node(&tokens, index);
+                args = make_args_node(&tokens, index)?;
                 continue;
             },
             token::TokenKind::TokenLessThan => {
-                func_ret_value_node = Some(parse_type::parse_type_node(&tokens, index).unwrap());
+                func_ret_value_node = Some(parse_type::parse_type_node(&tokens, index)?);
             }
             token::TokenKind::TokenFuncStart => {
                 func_start_keyword = true;
@@ -95,12 +120,12 @@ pub fn make_func_header(tokens: Vec<token::Token>, index: &mut usize) -> manager
         } else {
             func_ret_value_node.unwrap().value.clone()
         };
-        return manager::func::FuncNode {
+        return Ok(manager::func::FuncNode {
             name: tokens[func_name_index.unwrap()].lexeme.clone(),
             args: args,
             ret_type: type_api::change_txt_type_to_type(&type_name),
             nodes: Vec::<node::CalculNode>::new(),
-        };
+        });
     } else {
         panic!("syntax err: func define");
     }